+++ NEWS.md
@@ -2,6 +2,7 @@
 
 -   A new argument `deriv` added to `predict()` to calculate the first derivatives for `spl1D()` functions.
 -   Two examples in vignette updated with predictions of derivatives and corresponding standard errors.
+-   A bug in the `theta` argument of `LMMsolve()` fixed.
 
 # LMMsolver 1.0.10